+++ backend/app/src/main/java/eu/viandeendirect/repository/PackageTemplateRepository.java
@@ -3,5 +3,5 @@ package eu.viandeendirect.repository;
 import eu.viandeendirect.model.PackageTemplate;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PackageTemplateRepository extends CrudRepository<PackageTemplate, Long> {
+public interface PackageTemplateRepository extends CrudRepository<PackageTemplate, Integer> {
 }
